Cleanup r42210 "* (bug 16015) Change parsing of MediaWiki:Categorytree-member-num to allow &nbsp;"
'parsemag' was mysteriously chaned to 'parseinline', which is incorrect given that the result is HTML-escaped with 'escapenoenities' -- they don't play nicely together. :)

Changed it back to 'parsemag'. 'parseinline' *without* 'escapenoentities' might also seem like a good choice, but didn't seem to be what was intended.

Also broke the very very very long line for legibility.