Provide nice new arrows to replace bad old [+] [-] [x]
These folding arrows are based on the English Wikipedia setup, created
by Edokter. I don't see why all wikis can't benefit from them.

The only complication is with right-to-left languages, where ◄ is needed
instead of ►. I have updated the i18n file in a way that I hope is
sufficient.

When JavaScript is disabled, the fallback is a bit unintuitive, but no
worse than what is already there.

Bug: 36730
Change-Id: Ia15593ea5ee471c521820e2e69aa599fd4c4b921