Avoid href="#" for <a>
The href="#" shows a status line and allows to open a new tab on middle
click.

Add tabindex="0" and role="button" instead to still allow to focus the
element with the keyboard.

Add JavaScript event handler for the return key of the keyboard to still
allow to trigger the event with the keyboard.

Change-Id: I3f4c14d335f3a06bc67f148558e5b9a1534671bf