Avoid parsing and re-encoding ct-options in categoryTree.js
Follows-up 680cdd688ae1, which accessed the data attribute
using data() instead of attr(), making it thus be parsed as JSON,
only to be re-encoded with JSON.stringify().

I considered to instead change the 'wgCategoryTreePageCategoryOptions'
mw.config var to not be double-encoded as JSON, but concluded
it actually makes sense the way it is. Right now the options are,
from the client-side perspective, considered to be a blackbox string
set on either a link attribute, or in an mw.config key, and are
to be passed as-is (in whatever format) to the 'options' parameter
of the API module.

At least until we need these values on the client (which we don't right now)
it makes more sense not to have any js for decoding and re-coding.

Change-Id: I1e1e02a13d7a5293c9ae2e0ec61575ddb7a1f286