Make CategoryTree arrow expansion work outside mw-content-{ltr,rtl}
The old code relies on CSS Janus to transform the default case outside
mw-content-{ltr,rtl} which can't be applied here so let's consider
this case also by relying only on dir attributes, the same direction
as T371265.

In future we can use &:dir( ltr ) and &:dir( rtl ) as we do in new
codes but as that isn't supported on all browsers we support.

Change-Id: Icf6b8e184ce981e6e8f16dafa9e14a19154ff0e3