Apply conditions to correct column
Apparently even if we alias lt_title as cl_to in the select, it applies
the conditions set to the cl_to column. This switches the order of how
the tables are handled and together with the forced index cause a large
table scan resulting in a slow query.

Bug: T398823
Change-Id: I7677c913774d05fbdd87dedaefcc26aecf6ee9aa
(cherry picked from commit 27ba6c7569597b4ab04d36ebb8dd3c56e8382c76)